Migrate prototype map code to own widget with BLoC pattern
